Further transfer of the highscore methods to the new classes
