+++ src/main/java/lasersharks/Highscores.java
@@ -1,16 +1,37 @@
 package lasersharks;
 
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Scanner;
 
 import lasersharks.gui.LevelGUI;
 
 public class Highscores {
 
-  
   private static ArrayList<String> list = new ArrayList<String>();
   private static int score = LevelGUI.getScore();
+
+  /**
+   * Reads the current highscore list so it can be edited.
+   * 
+   * @return the current highscore list.
+   * @throws FileNotFoundException
+   */
+  public static ArrayList<String> readHighscore() throws FileNotFoundException {
+    @SuppressWarnings("resource")
+    Scanner sc = new Scanner(new File("highscores"));
+    ArrayList<String> list = new ArrayList<String>();
+    while (sc.hasNextLine()) {
+      list.add(sc.nextLine());
+    }
+
+    return list;
+
+  }
+
   /**
    * Method for writing the highscores. The highscores are saved in highscores.txt .
    * 
@@ -29,7 +50,7 @@ public class Highscores {
 
     try (FileWriter fw = new FileWriter("highscores")) {
       for (int i = 0; i < list.size(); i++) {
-        if (i < list.size() -1) {
+        if (i < list.size() - 1) {
           fw.write(list.get(i) + System.lineSeparator());
         } else {
           fw.write(list.get(i));
@@ -59,5 +80,18 @@ public class Highscores {
 
     return list;
   }
-  
+
+  /**
+   * Makes a nicely displayed string of the highscore
+   */
+
+  public String makeString() {
+    String res = "";
+    String li = System.lineSeparator();
+    for (int i = 0; i < list.size(); i++) {
+      res = res + list.get(i) + li;
+    }
+
+    return res;
+  }
 }
