+++ src/main/java/lasersharks/gui/LevelGUI.java
@@ -35,6 +35,7 @@ import javafx.stage.Stage;
 import lasersharks.Direction;
 import lasersharks.Fish;
 import lasersharks.Game;
+import lasersharks.Highscores;
 import lasersharks.LaserShark;
 import lasersharks.Position;
 import lasersharks.ScreenController;
@@ -258,8 +259,8 @@ public class LevelGUI extends Application {
    * @throws IOException
    */
   public void setWinSceneTrue() throws IOException {
-    list = readHighscore();
-    writeHighscore();
+    list = Highscores.readHighscore();
+    Highscores.writeHighscore();
     chooseWinScene = true;
     choosePlayScene = false;
     chooseLoseScene = false;
@@ -271,78 +272,11 @@ public class LevelGUI extends Application {
    * @throws IOException
    */
   public void setLoseSceneTrue() throws IOException {
-    list = readHighscore();
-    writeHighscore();
+    list = Highscores.readHighscore();
+    Highscores.writeHighscore();
     chooseWinScene = false;
     choosePlayScene = false;
     chooseLoseScene = true;
-  }
-
-  /**
-   * Reads the current highscore list so it can be edited.
-   * 
-   * @return the current highscore list.
-   * @throws FileNotFoundException
-   */
-  public static ArrayList<String> readHighscore() throws FileNotFoundException {
-    @SuppressWarnings("resource")
-    Scanner sc = new Scanner(new File("highscores"));
-    ArrayList<String> list = new ArrayList<String>();
-    while (sc.hasNextLine()) {
-      list.add(sc.nextLine());
-    }
-
-    return list;
-
-  }
-
-  /**
-   * Method for writing the highscores. The highscores are saved in highscores.txt .
-   * 
-   * @throws IOException
-   */
-  public static void writeHighscore() throws IOException {
-    for (int i = 0; i < list.size(); i++) {
-      if (score >= Integer.parseInt(list.get(i).substring(3))) {
-        list.remove(i);
-        list.add(i, i + ". " + score);
-        break;
-      }
-
-    }
-    fixHighscoreCount(list);
-
-    try (FileWriter fw = new FileWriter("highscores")) {
-      for (int i = 0; i < list.size(); i++) {
-        if (i < list.size() - 1) {
-          fw.write(list.get(i) + System.lineSeparator());
-        } else {
-          fw.write(list.get(i));
-        }
-      }
-
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-  }
-
-  /**
-   * Fixes the still erroneous list after the new score has been inserted in the writeHighscore()
-   * method. An example would be 1. 500 2. 250 2. 250 3. 100 4. 50.
-   * 
-   * @param list
-   * @return the correct highscore list
-   */
-  public static ArrayList<String> fixHighscoreCount(ArrayList<String> list) {
-    for (int i = 0; i < list.size(); i++) {
-      String newEntry = list.get(i);
-      list.remove(i);
-
-      newEntry = (i + 1) + ". " + newEntry.substring(3);
-      list.add(i, newEntry);
-    }
-
-    return list;
   }
 
   /**
